requirements.txt
# PBR Material Generation Training
torch>=2.1.0
torchvision>=0.16.0
accelerate>=0.25.0
transformers>=4.36.0
diffusers>=0.25.0
safetensors>=0.4.0

# Data
datasets>=2.16.0
Pillow>=10.0.0
albumentations>=1.3.0

# Monitoring
wandb>=0.16.0

# Config
pyyaml>=6.0
omegaconf>=2.3.0

# Utils
tqdm>=4.66.0
einops>=0.7.0
